+++ src/constants/src/AbstractConstants.php
@@ -37,7 +37,9 @@ abstract class AbstractConstants
 
         array_shift($arguments);
 
-        if ($result = self::translate($message, $arguments)) {
+        $result = self::translate($message, $arguments);
+        // If the result of translate doesn't exist, the result is equal with message, so we will skip it.
+        if ($result && $result !== $message) {
             return $result;
         }