+++ src/consul/src/Catalog.php
@@ -20,7 +20,7 @@ class Catalog extends Client implements CatalogInterface
             'body' => json_encode($node),
         ];
 
-        return $this->request('GET', '/v1/catalog/register', $params);
+        return $this->request('PUT', '/v1/catalog/register', $params);
     }
 
     public function deregister($node): ConsulResponse