+++ src/database/src/Model/Collection.php
@@ -18,9 +18,12 @@ use Hyperf\Utils\Arr;
 use Hyperf\Utils\Collection as BaseCollection;
 use Hyperf\Utils\Contracts\Arrayable;
 use Hyperf\Utils\Str;
+use Hyperf\Utils\Traits\Macroable;
 
 class Collection extends BaseCollection implements CompressInterface
 {
+    use Macroable;
+
     /**
      * Find a model in the collection by key.
      *
@@ -437,8 +440,9 @@ class Collection extends BaseCollection implements CompressInterface
 
     /**
      * Get a flattened array of the items in the collection.
+     * @param float|int $depth
      */
-    public function flatten(int $depth = INF): BaseCollection
+    public function flatten($depth = INF): BaseCollection
     {
         return $this->toBase()->flatten($depth);
     }