+++ src/guzzle/tests/Cases/CoroutineHandlerTest.php
@@ -18,6 +18,8 @@ use GuzzleHttp\Exception\RequestException;
 use GuzzleHttp\HandlerStack;
 use GuzzleHttp\Promise\PromiseInterface;
 use GuzzleHttp\Psr7\Request;
+use GuzzleHttp\RequestOptions;
+use GuzzleHttp\TransferStats;
 use Hyperf\Guzzle\CoroutineHandler;
 use HyperfTest\Guzzle\Stub\CoroutineHandlerStub;
 use PHPUnit\Framework\TestCase;
@@ -278,6 +280,36 @@ class CoroutineHandlerTest extends TestCase
         $ex = $handler->checkStatusCode($client, $request);
         $this->assertInstanceOf(RequestException::class, $ex);
         $this->assertSame('Server reset', $ex->getMessage());
+    }
+
+    public function testRequestOptionOnStats()
+    {
+        $url = 'http://127.0.0.1:9501';
+        $handler = new CoroutineHandlerStub();
+        $request = new Request('GET', $url . '/echo');
+
+        $bool = false;
+        $handler($request, [RequestOptions::ON_STATS => function (TransferStats $stats) use (&$bool) {
+            $bool = true;
+            $this->assertIsFloat($stats->getTransferTime());
+        }])->wait();
+        $this->assertTrue($bool);
+    }
+
+    public function testRequestOptionOnStatsInClient()
+    {
+        $bool = false;
+        $url = 'http://127.0.0.1:9501';
+        $client = new Client([
+            'handler' => new CoroutineHandlerStub(),
+            'base_uri' => $url,
+            RequestOptions::ON_STATS => function (TransferStats $stats) use (&$bool) {
+                $bool = true;
+                $this->assertIsFloat($stats->getTransferTime());
+            },
+        ]);
+        $client->get('/');
+        $this->assertTrue($bool);
     }
 
     protected function getHandler($options = [])