+++ src/json-rpc/tests/Stub/CalculatorServiceInterface.php
@@ -23,4 +23,8 @@ interface CalculatorServiceInterface
     public function array(int $a, int $b): array;
 
     public function error();
+
+    public function getString(): ?string;
+
+    public function callable(callable $a, ?callable $b): array ;
 }