+++ src/nsq/publish/nsq.php
@@ -12,6 +12,7 @@ declare(strict_types=1);
 
 return [
     'default' => [
+        'enable' => true,
         'host' => '127.0.0.1',
         'port' => 4150,
         'pool' => [